:bug: fixing bugs and improving

+++ burrows-wheeler/src/config.rs
@@ -1,2 +1,2 @@
 pub const WINDOW_SIZE: usize = 1024 * 1024;
-pub type Index = u64;
+pub type Index = u32;
